schedule weekly view now works properly!! move day of week sections to schedules comp. details:
So we moved a bunch of stuff from the WeeklyCalendarView to the Schedules.js. reason being,
it was much easier to update the state of day of week sections arrays when the arrow button
was clicked rather than checking if the index changed in the child component.
Next, we gotta style up, take into account mobile views, and filter unselected sections. :)