Routing with authentication now working, main pages setup.
Not focusing on styling just yet.  Had to make all components share loggedin state.
State of HOC withAuth is updated via methods passed down to components as props.
Whenever state is updated, components will see the changes.  Some pages, such as settings, are protected from non-loggedin users.
Next step is to display calendar in saved schedules component. One of my concerns is refresh tokens.  What if a particular user
is logged in on the mobile app and the web apps at the same time??  Well, logging in should generate a new token and before a user
tries to take a protected action, such as save schedule, the current token should be checked by the backend.  The backend should
accept one token from a user logged in to more than one platform, and if the token is not valid anymore, user should be logged out.