Saved schedules page functionality checkpoint. Open commit for details
Made some structural changes to Schedules.js.  Rather than having the mon, tue, wed, etc. arrays of divs in
the Schedules component, I removed them and just extract them whenever we need to using the schedules
prop that is passed in and the current index state to get a certain schedule from the passed down
schedules array.  Rather than setting state of those arrays, we just return an object that contains what we need when passing in the proper schedule.

For SavedSchedulesPage component, index changing of schedule is now handled.  Code is a little messy.
Want to refactor sometime but not soon.   Maybe after finals.